+++ src/space/core/connected_player/events.rs
@@ -7,41 +7,41 @@ use crate::space::core::{
 };
 
 pub struct NetSendServerTime {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetSendWorldEnvironment {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetUpdatePlayerCount {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct BoardingPlayer {
-    pub player_handle: u64,
+    pub player_handle: u32,
     pub player_character_name: String,
     pub entity: Entity,
 }
 
 pub struct InputExamineEntity {
-    pub handle: u64,
+    pub handle: u32,
     pub examine_entity_bits: u64,
     pub entity: Entity,
 }
 
 pub struct InputExamineMap {
-    pub handle: u64,
+    pub handle: u32,
     pub entity: Entity,
     pub gridmap_type: GridMapType,
     pub gridmap_cell_id: Vec3Int,
 }
 
 pub struct TextTreeInputSelection {
-    pub handle: u64,
+    pub handle: u32,
     pub menu_id: String,
     pub menu_selection: String,
     pub tab_action_id: String,
@@ -49,44 +49,44 @@ pub struct TextTreeInputSelection {
 }
 
 pub struct InputSceneReady {
-    pub handle: u64,
+    pub handle: u32,
     pub scene_type: String,
 }
 
 pub struct NetDoneBoarding {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetExamineEntity {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetOnBoarding {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetOnNewPlayerConnection {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetOnSetupUI {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct InputUIInputTransmitText {
-    pub handle: u64,
+    pub handle: u32,
     pub ui_type: String,
     pub node_path: String,
     pub input_text: String,
 }
 
 pub struct InputUIInput {
-    pub handle: u64,
+    pub handle: u32,
     pub node_class: UIInputNodeClass,
     pub action: UIInputAction,
     pub node_name: String,
@@ -147,7 +147,7 @@ pub struct InputMovementInput {
 }
 
 pub struct InputBuildGraphics {
-    pub handle: u64,
+    pub handle: u32,
 }
 
 #[derive(Debug)]
@@ -158,22 +158,22 @@ pub struct InputTabDataMap {
 }
 
 pub struct NetTabData {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetUIInputTransmitData {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetUserName {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetOnSpawning {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
