+++ src/space/core/map/events.rs
@@ -4,23 +4,23 @@ use bevy_math::Vec2;
 use crate::space::core::networking::resources::ReliableServerMessage;
 
 pub struct InputMapChangeDisplayMode {
-    pub handle: u64,
+    pub handle: u32,
     pub entity: Entity,
     pub display_mode: String,
 }
 
 pub struct InputMapRequestDisplayModes {
-    pub handle: u64,
+    pub handle: u32,
     pub entity: Entity,
 }
 
 pub struct NetRequestDisplayModes {
-    pub handle: u64,
+    pub handle: u32,
     pub message: ReliableServerMessage,
 }
 
 pub struct InputMap {
-    pub handle: u64,
+    pub handle: u32,
     pub entity: Entity,
     pub input: MapInput,
 }
